xtensa/esp32s3: Inspect if the MTD partition (factory/ota_0/ota_1) is mapped as text. Relocate the enum ota_img_ctrl_e and ota_img_bootseq_e to a directory visible to the application.
